requirements.txt
--index-url https://download.pytorch.org/whl/cu126
torch==2.7.1+cu126
torchvision==0.22.2+cu126
Pillow==11.0.0
numpy
matplotlib
opencv-python
scipy
tensorboardX
argparse
scikit-image
tqdm
imageio